5.11 added window.confirm, not tested yet

+++ part5/bloglist-frontend/src/App.jsx
@@ -36,7 +36,10 @@ const App = () => {
     }
   };
 
-  const handleDelete = async (id) => {
+  const handleDelete = async (id, title, author) => {
+    if (!window.confirm(`Remove Blog ${title} by ${author}?`)) {
+      return;
+    }
     try {
       const response = await blogService.remove(id);
       setIsError(false);
